+++ packages/core/src/types/index.ts
@@ -85,6 +85,8 @@ export type ShapeStyle = Partial<{
   fontSize: number;
 
   keepVisualSize: boolean;
+  // support more properties
+  [key: string]: any;
 }>;
 
 export interface IShapeBase extends IShape {
@@ -117,9 +119,11 @@ export type LoopConfig = Partial<{
 
 export interface LayoutConfig {
   type?: string;
+  gpuEnabled?: boolean;
   workerEnabled?: boolean;
   // 只有当 workerEnabled 为 true 时才生效
   workerScriptURL?: string;
+  onLayoutEnd?: () => void;
   [key: string]: unknown;
 }
 
@@ -181,10 +185,10 @@ export interface States {
 
 export interface StateStyles {
   [key: string]:
-    | ShapeStyle
-    | {
-        [key: string]: ShapeStyle;
-      };
+  | ShapeStyle
+  | {
+    [key: string]: ShapeStyle;
+  };
 }
 
 // model types (node edge group)
@@ -259,7 +263,7 @@ export interface GraphOptions {
     size: number | number[];
     color: string;
   }> &
-    ModelStyle;
+  ModelStyle;
 
   /**
    * 默认状态下边的配置，比如 type, size, color。会被写入的 data 覆盖。
@@ -269,7 +273,7 @@ export interface GraphOptions {
     size: number | number[];
     color: string;
   }> &
-    ModelStyle;
+  ModelStyle;
 
   /**
    * Combo 默认配置
@@ -279,7 +283,7 @@ export interface GraphOptions {
     size: number | number[];
     color: string;
   }> &
-    ModelStyle;
+  ModelStyle;
 
   nodeStateStyles?: StateStyles;
 
@@ -409,10 +413,10 @@ export interface TreeGraphData {
   depth?: number;
   collapsed?: boolean;
   style?:
-    | ShapeStyle
-    | {
-        [key: string]: ShapeStyle;
-      };
+  | ShapeStyle
+  | {
+    [key: string]: ShapeStyle;
+  };
   stateStyles?: StateStyles;
   [key: string]: unknown;
 }
@@ -550,6 +554,7 @@ export interface GraphData {
   nodes?: NodeConfig[];
   edges?: EdgeConfig[];
   combos?: ComboConfig[];
+  [key: string]: any;
 }
 
 export interface GraphAnimateConfig extends AnimateCfg {