+++ packages/pc/tests/unit/layout/circular-web-worker-spec.ts
@@ -1,4 +1,3 @@
-// 注意：这里不能直接require原始的src文件，而要使用build后的文件，因为web worker代码是通过worker-loader内联进来的
 import G6 from '../../../src';
 import dataset from './data';
 import { mathEqual } from './util';
@@ -23,6 +22,8 @@ describe('circular layout(web worker)', () => {
       defaultNode: { size: 10 },
     });
     graph.data(data);
+    graph.render();
+
     graph.on('afterlayout', () => {
       const width = graph.get('width');
       const height = graph.get('height');
@@ -34,7 +35,6 @@ describe('circular layout(web worker)', () => {
       graph.destroy();
       done();
     });
-    graph.render();
   });
 
   it('circular(web worker) counterclockwise, and fixed radius, start angle, end angle', done => {
@@ -54,6 +54,8 @@ describe('circular layout(web worker)', () => {
       defaultNode: { size: 10 },
     });
     graph.data(data);
+    graph.render();
+
     graph.on('afterlayout', () => {
       const pos = (200 * Math.sqrt(2)) / 2;
       expect(mathEqual(data.nodes[0].x, 250 + pos)).toEqual(true);
@@ -61,6 +63,5 @@ describe('circular layout(web worker)', () => {
       graph.destroy();
       done();
     });
-    graph.render();
   });
 });